reactive_graph_runtime_web_api::web_resource_manager

Trait WebResourceManager

pub trait WebResourceManager:
    Send
    + Sync
    + Lifecycle {
    // Required methods
    fn has(&self, context_path: String) -> bool;
    fn get(&self, context_path: String) -> Option<Arc<dyn WebResourceProvider>>;
    fn get_default(&self) -> Option<Arc<dyn WebResourceProvider>>;
    fn register_provider<'life0, 'async_trait>(
        &'life0 self,
        web_resource_provider: Arc<dyn WebResourceProvider>,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn unregister_provider<'life0, 'async_trait>(
        &'life0 self,
        id: Uuid,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn has(&self, context_path: String) -> bool

Returns true, if a web resource provider exists with the given context path.

fn get(&self, context_path: String) -> Option<Arc<dyn WebResourceProvider>>

Returns the web resource provider with the given context path.

fn get_default(&self) -> Option<Arc<dyn WebResourceProvider>>

Returns the default web resource provider.

fn register_provider<'life0, 'async_trait>( &'life0 self, web_resource_provider: Arc<dyn WebResourceProvider>, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Registers a web resource provider.

fn unregister_provider<'life0, 'async_trait>( &'life0 self, id: Uuid, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Unregisters a web resource provider.